Move content review components into 'review' app
