+++ orb/urls.py
@@ -62,6 +62,7 @@ urlpatterns = patterns('',
     url(r'^resource/', include('orb.resources.urls')),
     url(r'^resource/rate/', include('orb.rating.urls')),
     url(r'^resource/bookmark/', include('orb.bookmark.urls')),
+    url(r'^review/', include('orb.review.urls')),
     url(r'^viz/', include('orb.viz.urls')),
     url(r'^toolkits/', include('orb.toolkits.urls')),
     url(r'^i18n/', include('django.conf.urls.i18n')),
